+++ packages/SwingSet/src/kernel/vatManager/lockdown-subprocess-xsnap.js
@@ -1,1 +1,1 @@
-import '@agoric/xsnap/src/bootstrap';
+import '@agoric/xsnap/lib/bootstrap';
