+++ packages/xsnap/lib/bootstrap.js
