+++ packages/xsnap/lib/console-shim.js
