+++ packages/xsnap/lib/lockdown-shim.js
