+++ packages/xsnap/lib/text-shim.js
