+++ packages/xsnap/package.json
@@ -18,9 +18,9 @@
     "build": "yarn build:bin && yarn build:bundle",
     "clean": "rm -rf build",
     "lint": "yarn lint:js && yarn lint:types",
-    "lint:js": "eslint 'src/**/*.js'",
+    "lint:js": "eslint 'src/**/*.js' 'lib/**/*.js'",
     "lint:types": "tsc -p jsconfig.json",
-    "lint-fix": "eslint --fix 'src/**/*.js'",
+    "lint-fix": "eslint --fix 'src/**/*.js' 'lib/**/*.js'",
     "lint-check": "yarn lint",
     "test": "ava",
     "postinstall": "yarn build"
